DecodeIR.h: change frequency_t to uint32_t
to be compatible with current Infrared4Arduino.

+++ src/DecodeIR.h
@@ -32,7 +32,7 @@
 typedef unsigned frequency_t;
 typedef unsigned microseconds_t;
 #else
-typedef uint16_t frequency_t;
+typedef uint32_t frequency_t;
 typedef uint16_t microseconds_t;
 #endif
